User

+++ application/migrations/001_create_posts.php
@@ -7,7 +7,11 @@ class Migration_Create_posts extends CI_Migration
     public function up()
     {
         $this->dbforge->add_field([
-            'id' => [
+            'user_id' => [
+                'type' => 'INT',
+                'unsigned' => TRUE
+            ],
+            'post_id' => [
                 'type' => 'INT',
                 'auto_increment' => TRUE
             ],
@@ -18,16 +22,22 @@ class Migration_Create_posts extends CI_Migration
             'content' => [
                 'type' => 'TEXT'
             ],
+            'tag_id' => [
+                'type'     => 'INT',
+                'unsigned' => TRUE
+            ],
             'created_at' => [
                 'type' => 'DATETIME'
             ]
         ]);
-        $this->dbforge->add_key('id', TRUE);
+        $this->dbforge->add_key('post_id', TRUE);
         $this->dbforge->create_table('posts');
     }
 
     public function down()
     {
-        $this->dbforge->drop_table('posts');
+        if ($this->db->table_exists('posts')) {
+            $this->dbforge->drop_table('posts');
+        }
     }
 }
